crudApi/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
